Added Bookshelf and Cat Tower
I had purchased an asset pack with furniture in it but noticed there were things missing that I would have liked to add into my game. So, I'm working on creating other assets to match the style of the asset pack and implementing them into the background. For now, I have added the bookshelf and the cat tower. There are a few more things that I am working to add!