Updated Info Page, Added Second Level
Updated the info page to instead be a pop up when clicking the info button, as to not reset the layout when looking at the controls. Added a few items and points so that the door opens and unlocks the second level. Added some items to knock over in the second level. Total points at this moment is 40, though I'd like to keep adding to this.